Revert "Extract public ES2K functions (#475)"
This reverts commit b6db3a4cedf53184e862d7d44ba8e20c129ecc52.
