+++ ovs-p4rt/sidecar/es2k/CMakeLists.txt
@@ -6,6 +6,5 @@
 target_sources(ovs_sidecar_o PRIVATE
     ovsp4rt_es2k_private.cc
     ovsp4rt_es2k_private.h
-    ovsp4rt_es2k_public.cc
     p4_name_mapping.h
 )
